Move testing code into new directory
Updates `.gitignore` and splits some js out of index.js
